requirements.txt
gradio
google-generativeai
google-genai
dotenv